+++ js/gtUINodes.js
@@ -1,4 +1,7 @@
 import { app } from "../../../scripts/app.js";
+import { ComfyButtonGroup } from "../../scripts/ui/components/buttonGroup.js";
+import { ComfyButton } from "../../scripts/ui/components/button.js";
+
 import { api } from "../../../scripts/api.js";
 import { nodeFixes } from "./nodeFixes.js";
 import { setupConfigurationNodes } from "./ConfigurationNodes.js";
@@ -12,8 +15,20 @@ import { setupMenuSeparator } from "./gtUIMenuSeparator.js";
 import { keys_organized } from "./griptape_api_keys.js";
 import { setupVisibilityToggles } from "./NodesWithVisibilityToggles.js";
 import { setupCodeExecutionNode } from "./CodeExecutionNode.js";  
+
+function addGriptapeTopBarButtons() {
+  const buttons = [];
+  const griptapeButton = new ComfyButton({
+    tooltip: "Griptape",
+    app,
+    enabled: true,
+    classList: "comfyui-button comfyui-menu-mobile-collapse primary",
+});
+  console.log(griptapeButton);
+}
 app.registerExtension({
   name: "comfy.gtUI",
+  addGriptapeTopBarButtons,
   beforeConfigureGraph: (graphData, missingNodeTypes) => {
     for (let node of graphData.nodes) {
       if (nodeFixes.fixes[node.type]) {
@@ -35,13 +50,25 @@ app.registerExtension({
       name: "default_config",
       type: "dict",
       defaultValue: "",
+      tooltip: "To set this, use the Griptape: Set Default Agent node.",
+    });
+    app.ui.settings.addSetting({
+      id: `Griptape.allow_code_execution_dangerous`,
+      category: ["Griptape", "!Griptape", "code_execution_dangerous"],
+      name: "Enable Insecure Code Execution [DANGER]",
+      type: "boolean",
+      tooltip: "When enabled, the Griptape Util: Code Execution node not check for dangerous code.\n\n[WARNING] This setting is dangerous and should only be enabled if you know what you are doing.",
+      defaultValue: false,
+      onChange: (newVal, oldVal) => { if (newVal == true) { console.warn("Griptape Utils: Dangerous Code Execution enabled: ", newVal)} },
     });
     app.ui.settings.addSetting({
       id: `Griptape.allow_code_execution`,
-      category: ["Griptape", "Griptape", "code_execution"],
+      category: ["Griptape", "!Griptape", "code_execution"],
       name: "Enable Code Execution Nodes",
       type: "boolean",
+      tooltip: "When enabled, the Griptape Util: Code Execution node will be available for use.",
       defaultValue: false,
+      onChange: (newVal, oldVal) => { console.log("Setting got changed!", newVal) },
     });
     Object.entries(keys_organized).forEach(([category, keys]) => {
       keys.forEach((key) => {